Add node + mocha
