+++ package.json
@@ -3,14 +3,19 @@
   "scripts": {
     "start": "firefox-proxy & ./node_modules/.bin/webpack --watch & ./node_modules/.bin/static --port 8000",
     "lint-css": "stylelint js/components/*.css",
-    "lint-js": "eslint js"
+    "lint-js": "eslint js",
+    "test": "./bin/test"
   },
   "dependencies": {
+    "babel-cli": "^6.7.5",
+    "babel-polyfill": "^6.7.4",
+    "babel-preset-stage-0": "^6.5.0",
     "co": "=4.6.0",
     "codemirror": "^5.1.0",
     "ff-devtools-libs": "^0.1.1",
     "immutable": "^3.7.6",
     "json-loader": "^0.5.4",
+    "mocha": "^2.4.5",
     "net": "^1.0.2",
     "node-static": "^0.7.7",
     "react": "=0.14.7",
@@ -25,13 +30,17 @@
     "ws": "^1.0.1"
   },
   "devDependencies": {
+    "babel": "^6.5.2",
+    "babel-core": "^6.7.6",
+    "babel-loader": "^6.2.4",
+    "babel-polyfill": "^6.7.4",
+    "babel-preset-es2015": "^6.6.0",
     "css-loader": "^0.23.1",
     "eslint": "^2.7.0",
     "eslint-plugin-mozilla": "0.0.3",
     "eslint-plugin-react": "^4.3.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "rimraf": "^2.5.2",
-    "style-loader": "^0.13.1",
-    "xpcshell-unit-test": "0.0.2"
+    "style-loader": "^0.13.1"
   }
 }
